+++ node-zcash.d.ts
@@ -1,0 +1,3 @@
+import type { ZcashIo } from './src/zcash/zcashIo'
+
+export function makeZcashIo(opts: { documentDirectory: string }): ZcashIo