+++ src/zcash/ZcashTools.ts
@@ -12,7 +12,6 @@ import {
   EdgeWalletInfo,
   JsonObject
 } from 'edge-core-js/types'
-import { Tools as ToolsType } from 'react-native-zcash'
 import { base64url } from 'rfc4648'
 
 import { PluginEnvironment } from '../common/innerPlugin'
@@ -42,7 +41,7 @@ export class ZcashTools implements EdgeCurrencyTools {
   currencyInfo: EdgeCurrencyInfo
   io: EdgeIo
   networkInfo: ZcashNetworkInfo
-  nativeTools: typeof ToolsType
+  nativeTools: ZcashIo['Tools']
 
   constructor(env: PluginEnvironment<ZcashNetworkInfo>) {
     const { builtinTokens, currencyInfo, io, networkInfo } = env